mybatis 의존성 추가 및 apt_seq로 아파트 정보 조회 기능 작업
