+++ src/main/java/com/ssafy/home/Heo/common/exception/BaseException.java
@@ -1,0 +1,14 @@
+package com.ssafy.home.Heo.common.exception;
+
+import com.ssafy.home.Heo.common.base.BaseResponseStatus;
+import lombok.Getter;
+
+//@Getter
+//public class BaseException extends RuntimeException{
+//
+//    private final BaseResponseStatus status;
+//
+//    public BaseException(BaseResponseStatus status) {
+//        this.status = status;
+//    }
+//}
