+++ src/main/java/com/ssafy/home/Heo/common/exception/BaseExceptionHandler.java
@@ -1,0 +1,22 @@
+package com.ssafy.home.Heo.common.exception;
+
+import com.ssafy.home.Heo.common.base.BaseResponse;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.ExceptionHandler;
+import org.springframework.web.bind.annotation.RestControllerAdvice;
+
+//@RestControllerAdvice
+//@Slf4j
+//public class BaseExceptionHandler {
+//
+//    /**
+//     * 발생한 예외 처리
+//     */
+//    @ExceptionHandler(BaseException.class)
+//    protected ResponseEntity<BaseResponse<Void>> BaseError(BaseException e) {
+//        BaseResponse<Void> response = new BaseResponse<>(e.getStatus());
+//        log.error("BaseException -> {}({})", e.getStatus(), e.getStatus().getMessage(), e);
+//        return new ResponseEntity<>(response, response.httpStatus());
+//    }
+//}
