+++ src/main/resources/mappers/house.xml
@@ -2,19 +2,29 @@
 <!DOCTYPE mapper
   PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
   "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.ssafy.home.Heo.house.model.dao.UserDao">
+<mapper namespace="com.ssafy.home.Heo.house.repository.HouseDao">
 
 	
 	<!-- 아이디, 비번을 통한 유저정보 조회(id, name 데이터를 조회) -->
-	<select id="findByIdAndPassword" parameterType="User">
-		select id, name, email, role
-		  from users
-		 where id = #{id} and password = #{password}
+	<select id="findHouseByAptSeq" parameterType="String"   resultType="com.ssafy.home.Heo.house.dto.out.HouseDetailResponseDto">
+		SELECT
+			apt_seq,
+			sgg_cd,
+			umd_cd,
+			umd_nm,
+			jibun,
+			road_nm_sgg_cd,
+			road_nm,
+			road_nm_bonbun,
+			road_nm_bubun,
+			apt_nm,
+			build_year,
+			latitude,
+			longitude
+		FROM houseinfos
+		WHERE apt_seq = #{aptSeq}
 	</select>
 	
-	<insert id="join" parameterType="User">
-		insert into users
-		values(#{id}, #{password}, #{name}, #{email}, #{role})
-	</insert>
+
 	
 </mapper>
