--- src/main/resources/mappers/notebook.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.ssafy.offline.model.dao.NotebookDao">
-
-	<!-- 노트북 목록 조회 -->
-	<select id="selectAll" resultMap="notebookMap">
-		select notebook_id, notebook_name
-		 from notebook
-	</select>
-
-	<resultMap type="Notebook" id="notebookMap">
-		<id column="notebook_id" property="notebookId"/>
-		<result column="notebook_name" property="notebookName"/>
-	</resultMap>
-	
-</mapper>
