--- src/main/resources/mappers/rental.xml
@@ -1,63 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.ssafy.offline.model.dao.RentalDao">
-
-	<!-- 노트북 대여하기 -->
-	<insert id="addRental" parameterType="Rental">
-		insert into rental(user_id, notebook_id, start_time, end_time)
-		values(#{userId}, #{notebookId}, #{startTime}, #{endTime})	
-	</insert>
-	
-	<select id="selectById" parameterType="String" resultMap="rentalNotebookMap">
-		select r.rental_id, n.notebook_name, r.start_time, r.end_time 
-		  from rental r 
-		  join notebook n on r.notebook_id = n.notebook_id 
-		 where user_id = #{id}; 
-	</select>
-	
-	<delete id="delete" parameterType="List">
-		delete from rental
-		<foreach item="rentalId" collection="rentalId" open="where rental_id IN(" close=")" separator=",">
-			#{rentalId}
-		</foreach> 
-	</delete>
-	
-	<select id="selectAll" parameterType="SearchCondition" resultMap="rentalNotebookMap">
-		select r.rental_id, r.user_id, n.notebook_name, r.start_time, r.end_time
-		  from rental r 
-		  join notebook n on r.notebook_id = n.notebook_id
-		  <where>
-		  	<if test="date != null">
-				and r.start_time >= #{date}
-			</if>
-			<if test="notebookId != 0">
-				and n.notebook_id = #{notebookId}
-			</if>
-		  </where> 
-		  order by rental_id
-	</select>
-	
-	<delete id="deleteById" parameterType="int">
-		delete from rental
- 		 where rental_id = #{rentalId}
-	</delete>
-
-	<resultMap type="Rental" id="rentalMap">
-		<id column="rental_id" property="rentalId"/>
-		<result column="user_id" property="userId"/>
-		<result column="notebook_id" property="notebookId"/>
-		<result column="start_time" property="startTime"/>
-		<result column="end_time" property="endTime"/>
-	</resultMap>
-	
-	<resultMap type="Notebook" id="notebookMap">
-		<id column="notebook_id" property="notebookId"/>
-		<result column="notebook_name" property="notebookName"/>
-	</resultMap>
-	
-	<resultMap type="rental" id="rentalNotebookMap" extends="rentalMap">
-        <association property="notebook" column="notebook_id" resultMap="notebookMap"></association>
-    </resultMap>
-</mapper>
