--- src/main/resources/mappers/user.xml
@@ -1,20 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-  PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-  "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.ssafy.offline.model.dao.UserDao">
-
-	
-	<!-- 아이디, 비번을 통한 유저정보 조회(id, name 데이터를 조회) -->
-	<select id="findByIdAndPassword" parameterType="User">
-		select id, name, email, role
-		  from users
-		 where id = #{id} and password = #{password}
-	</select>
-	
-	<insert id="join" parameterType="User">
-		insert into users
-		values(#{id}, #{password}, #{name}, #{email}, #{role})
-	</insert>
-	
-</mapper>
